move to spark-sql
